requirements.txt
pytorch
transformers
datasets
cudatoolkit=11.0
matplotlib
numpy